requirements.txt
aiohttp==3.5.4
altgraph==0.16.1
apparmor==2.13.2
appdirs==1.4.3
application-utility==1.1.0
asn1crypto==0.24.0
astroid==2.2.5
async-timeout==3.0.1
attrs==19.1.0
autopep8==1.4.4
awscli==1.16.140
beautifulsoup4==4.7.1
boto3==1.9.130
botocore==1.12.130
btrfsutil==1.1.0
CacheControl==0.12.5
cffi==1.12.2
chardet==3.0.4
chrome-gnome-shell==0.0.0
colorama==0.3.9
coloredlogs==10.0
cryptography==2.6.1
cupshelpers==1.0
discord.py==1.0.1
distlib==0.2.8
distro==1.4.0
docutils==0.14
entrypoints==0.3
flake8==3.7.7
future==0.17.1
gufw==18.10.0
html5lib==1.0.1
humanfriendly==4.17
idna==2.8
isort==4.3.16
Jade-Application-Kit==1.3.5
jmespath==0.9.4
lazy-object-proxy==1.3.1
LibAppArmor==2.13.2
lockfile==0.12.2
louis==3.9.0
macholib==1.11
mccabe==0.6.1
msgpack==0.6.1
multidict==4.5.2
npyscreen==4.10.5
oauthlib==3.0.1
packaging==19.0
pacman-mirrors==4.14.1
paypalrestsdk==1.13.1
pefile==2018.8.8
pep517==0.5.0
pep8==1.7.1
petname==2.0
Pillow==6.0.0
ply==3.11
progress==1.5
pwquality==1.4.0
pyasn1==0.4.5
PyAutoGUI==0.9.42
pycairo==1.18.0
pycodestyle==2.5.0
pycparser==2.19
pycups==1.9.74
pycurl==7.43.0.2
pyflakes==2.1.1
PyGetWindow==0.0.4
PyGObject==3.32.0
PyInstaller==3.4
pylint==2.3.1
PyMsgBox==1.0.6
PyNaCl==1.2.1
pyOpenSSL==19.0.0
pyparsing==2.3.1
PyQt5==5.12.1
PyQt5-sip==4.19.16
PyRect==0.1.4
PyScreeze==0.1.20
python-apt==0.0.0
python-dateutil==2.8.0
python-debian==0.1.34+git20190401
python-distutils-extra==2.39
pytoml==0.1.20
PyTweening==1.0.3
pyxdg==0.26
PyYAML==3.13
requests==2.21.0
requests-oauthlib==1.2.0
retrying==1.3.3
rsa==3.4.2
s3transfer==0.2.0
six==1.12.0
soupsieve==1.9
ssoclient==2.1.1
team==1.0
typed-ast==1.3.1
ufw==0.36
urllib3==1.24.1
verboselogs==1.7
webencodings==0.5.1
websockets==6.0
wrapt==1.11.1
yarl==1.3.0